+++ angular/scripts/route/homeRoute.js
@@ -173,8 +173,13 @@ angular.module('theOneBlog')
       }
     };
   }])
-  .config(['$stateProvider', '$urlRouterProvider',function($stateProvider, $urlRouterProvider) {
+  .config(['$stateProvider', '$urlRouterProvider', '$locationProvider', function($stateProvider, $urlRouterProvider ,$locationProvider) {
+      $locationProvider.html5Mode({
+        enabled: true,
+        requireBase: false
+      });
       $urlRouterProvider.otherwise('/');
+
       $stateProvider
         .state('main',{
           url:'/',